refactor: Move pipeline crates to pipelines module
